refactor(eg): separate eg.Agent from eg (#327)
Close #323
