+++ config/dependency.js
@@ -1,6 +1,8 @@
 module.exports = {
+	eg: [
+		"src/module.js"
+	],
 	egCore : [
-		"src/module.js",
 		"src/eg.js",
 		"src/class.js",
 		"src/component.js"
@@ -11,7 +13,6 @@ module.exports = {
 		"src/hook/transform.js",
 		"src/hook/cssPrefix.js",
 		"src/hook/pauseResume.js",
-		"src/plugin/apply.js",
 		"src/plugin/persist.js"
 	],
 	hammer : [
