+++ test/unit/agent.test.html
@@ -7,47 +7,32 @@
 <link rel="stylesheet" href="../../bower_components/qunit/qunit/qunit.css">
 <!-- user css start -->
 <style>
-.box {
-	width:120px;
-	height:120px;
-	position:absolute;
-	border:solid;
-	padding: 0px 0px;
-	color:blue;
-}
-
-.otherStyleBox {
-	position:absolute;
-	border:solid 1px;
-	left:50px;
-	height:50px;
-}
 </style>
 <!-- user css end -->
 <script src="../../bower_components/qunit/qunit/qunit.js"></script>
 <script src="../../bower_components/jquery/jquery.js"></script>
-<!-- <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/1.12.1/jquery.min.js"></script> -->
 </head>
 <body>
- <h1 id="qunit-header">Pause/Resume Test</h1>
+ <h1 id="qunit-header">Extend Test</h1>
  <h2 id="qunit-banner"></h2>
  <div id="qunit-testrunner-toolbar"></div>
  <h2 id="qunit-userAgent"></h2>
  <ol id="qunit-tests"></ol>
- <div id="qunit-fixture1">
+ <div id="qunit-fixture">
 <!-- UI content Start -->
- <div id="box1" class="box"></div>
+ test markup, will be hidden
  <!-- UI content End -->
  </div>
+
 
 <!-- user script files start -->
 <script src="lib/module.js"></script>
 <script src="../../src/agent.js"></script>
 <script src="../../src/eg.js"></script>
-<script src="../../src/hook/pauseResume.js"></script>
 <!-- user script files end -->
 <!-- test script files start -->
-<script src="js/pauseResume.test.js"></script>
+<script src="js/agent_info.js"></script>
+<script src="js/agent.test.js"></script>
 <!-- test script files end -->
 </body>
 </html>
