+++ test/unit/class.test.html
@@ -26,6 +26,7 @@
 
 <!-- user script files start -->
 <script src="lib/module.js"></script>
+<script src="../../src/agent.js"></script>
 <script src="../../src/eg.js"></script>
 <script src="../../src/class.js"></script>
 <!-- user script files end -->
