+++ test/unit/eg.test.html
@@ -27,9 +27,11 @@
 
 <!-- user script files start -->
 <script src="lib/module.js"></script>
+<script src="../../src/agent.js"></script>
 <script src="../../src/eg.js"></script>
 <!-- user script files end -->
 <!-- test script files start -->
+<script src="js/agent_info.js"></script>
 <script src="js/eg.test.js"></script>
 <!-- test script files end -->
 </body>
