+++ test/unit/flicking.test.html
@@ -98,6 +98,7 @@
 	Simulator.events.touch.fakeSupport();
 </script>
 <script src="lib/module.js"></script>
+<script src="../../src/agent.js"></script>
 <script src="../../src/eg.js"></script>
 <script src="../../src/class.js"></script>
 <script src="../../src/component.js"></script>
