+++ test/unit/module.test.html
@@ -35,6 +35,7 @@ var eg = {
 };
 </script>
 <script src="../../src/module.js"></script>
+<script src="../../src/agent.js"></script>
 <script src="../../src/eg.js"></script>
 <script src="../../src/class.js"></script>
 <script src="../../src/component.js"></script>
