+++ test/unit/movableCoord.test.html
@@ -48,6 +48,7 @@
     Simulator.events.touch.fakeSupport();
 </script>
 <script src="lib/module.js"></script>
+<script src="../../src/agent.js"></script>
 <script src="../../src/eg.js"></script>
 <script src="../../src/class.js"></script>
 <script src="../../src/component.js"></script>
