+++ test/unit/persist.test.html
@@ -24,6 +24,7 @@
  <!-- UI content End -->
  </div>
 <script src="lib/module.js"></script>
+<script src="../../src/agent.js"></script>
 <script src="../../src/eg.js"></script>
 <script type="text/javascript">
 eg.debug = true;
