+++ test/unit/rotate.test.html
@@ -27,6 +27,7 @@
 <!-- user script files start -->
 <script src="lib/sinon-1.14.1.js"></script>
 <script src="lib/module.js"></script>
+<script src="../../src/agent.js"></script>
 <script src="../../src/eg.js"></script>
 <script src="../../src/customEvent/rotate.js"></script>
 <!-- user script files end -->
