+++ test/unit/scrollEnd.test.html
@@ -26,6 +26,7 @@
 
 <!-- user script files start -->
 <script src="lib/module.js"></script>
+ <script src="../../src/agent.js"></script>
  <script src="../../src/eg.js"></script>
 <script src="../../src/customEvent/scrollEnd.js"></script>
 <script src="js/scrollEnd.test.js"></script>
