+++ test/unit/transform.test.html
@@ -33,6 +33,7 @@
 
 <!-- user script files start -->
 <script src="lib/module.js"></script>
+<script src="../../src/agent.js"></script>
 <script src="../../src/eg.js"></script>
 <script src="../../src/class.js"></script>
 <script src="../../src/hook/cssPrefix.js"></script>
