+++ test/unit/visible.test.html
@@ -263,6 +263,7 @@ same work.
  </div>
 <!-- user script files start -->
 <script src="lib/module.js"></script>
+<script src="../../src/agent.js"></script>
 <script src="../../src/eg.js"></script>
 <script src="../../src/class.js"></script>
 <script src="../../src/component.js"></script>
